requirements.txt
fastapi>=0.95.0
uvicorn>=0.22.0
pydantic>=2.0.0
llmlingua>=0.2.2
tiktoken>=0.4.0
python-dotenv>=1.0.0
requests>=2.28.0
PyYAML>=6.0